commented out genres code
the code and function for retrieving the genre of an artist was suddenly not working? I ended up commenting it out just to get basic datasets for 500, 1000, and 1500 songs. These also do not incorporate temporal aspects. I can try to troubleshoot the genres a later time, its time to move on to the recommender